+++ public/js/user.js
@@ -9,6 +9,8 @@ $(document).ready(function() {
   var $userZip = $("#zip");
   var $submitBtn = $("#submit");
   var $userList = $("#user-list");
+  var $modal = $("#input-modal");
+  var $modalPara = $(".modal-content > p");
 
   // The API object contains methods for each kind of request we'll make
   var API = {
@@ -30,45 +32,65 @@ $(document).ready(function() {
     },
     deleteUser: function(id) {
       return $.ajax({
-        url: "api/users/" + id,
+        url: "/api/users/" + id,
         type: "DELETE"
       });
     }
-  };
-
-  // refreshUsers gets new users from the db and repopulates the list
-  var refreshUsers = function() {
-    API.getUsers().then(function(data) {
-      var $users = data.map(function(user) {
-        var $a = $("<a>")
-          .text(user.text)
-          .attr("href", "/user/" + user.id);
-
-        var $li = $("<li>")
-          .attr({
-            class: "list-group-item",
-            "data-id": user.id
-          })
-          .append($a);
-
-        var $button = $("<button>")
-          .addClass("btn btn-danger float-right delete")
-          .text("ｘ");
-
-        $li.append($button);
-
-        return $li;
-      });
-
-      $userList.empty();
-      $userList.append($users);
-    });
   };
 
   // handleFormSubmit is called whenever we submit a new user
   // Save the new user to the db and refresh the list
   var handleFormSubmit = function(event) {
     event.preventDefault();
+
+    // User name validation
+    if (!$userName.val()) {
+      $modalPara.text("User name cannot be blank.");
+      $modal.modal("open");
+      return;
+    }
+
+    // User name validation
+    if (
+      !$userEmail.val() ||
+      !$userEmail
+        .val()
+        .match(
+          /^([a-zA-Z0-9_\-\.]+)@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.)|(([a-zA-Z0-9\-]+\.)+))([a-zA-Z]{2,4}|[0-9]{1,3})(\]?)$/
+        )
+    ) {
+      $modalPara.text("Please enter a valid email address.");
+      $modal.modal("open");
+      return;
+    }
+
+    // User phone validation
+    if (!$userPhone.val()) {
+      $modalPara.text("Phone number cannot be blank.");
+      $modal.modal("open");
+      return;
+    }
+
+    // User city validation
+    if (!$userCity.val()) {
+      $modalPara.text("City cannot be blank.");
+      $modal.modal("open");
+      return;
+    }
+
+    // User state validation
+    if (!$userState.val()) {
+      $modalPara.text("You must choose a state.");
+      $modal.modal("open");
+      return;
+    }
+
+    // User zip code validation
+    if (!$userZip.val() || !$userZip.val().match(/^\d{5}$/)) {
+      $modalPara.text("Zip code cannot be blank and must contain 5 digits.");
+      $modal.modal("open");
+      return;
+    }
 
     var user = {
       name: $userName.val().trim(),
@@ -79,15 +101,9 @@ $(document).ready(function() {
       zip: $userZip.val().trim()
     };
 
-    // TODO Validation
-    if (!(user.name && user.email)) {
-      alert("You must enter a user name and email!");
-      return;
-    }
-
     API.saveUser(user).then(function() {
       alert("User saved!");
-      refreshUsers();
+      location.assign("/all-users");
     });
 
     $userName.val("");
@@ -101,13 +117,14 @@ $(document).ready(function() {
   // handleDeleteBtnClick is called when a user's delete button is clicked
   // Remove the user from the db and refresh the list
   var handleDeleteBtnClick = function() {
-    var idToDelete = $(this)
-      .parent()
-      .attr("data-id");
-
-    API.deleteUser(idToDelete).then(function() {
-      refreshUsers();
-    });
+    if (confirm("Are you sure you want to delete this user?")) {
+      var idToDelete = $(this)
+        .parent()
+        .attr("data-id");
+      API.deleteUser(idToDelete).then(function() {
+        location.reload();
+      });
+    }
   };
 
   // Add event listeners to the submit and delete buttons
